Move board manager classes to management package
Refactor package layout by moving BoardManager, HeadlessBoardManager and ThreadActionListener from app.freerouting.interactive to app.freerouting.management. Update imports across main code and tests to reference the new package, and adjust related usages (e.g. JobControllerV1, GuiManager, Autorouter/threads, RoutingJobScheduler, BoardLoader, GuiBoardManager). Change BoardManager.getInteractiveSettings() return type from InteractiveSettings to GuiSettings and ensure HeadlessBoardManager.getInteractiveSettings() returns GuiSettings (still null in headless mode). Update deprecated get_settings() to delegate to the new type. Minor formatting and javadoc tweaks and clean up of archunit_store entries to reflect the moves.
