+++ src/main/java/app/freerouting/interactive/InteractiveActionThread.java
@@ -9,6 +9,7 @@ import app.freerouting.settings.GlobalSettings;
 import java.awt.Graphics;
 import java.io.IOException;
 import java.util.ArrayList;
+import app.freerouting.management.ThreadActionListener;
 import java.util.List;
 
 /**
