+++ src/main/java/app/freerouting/management/BoardLoader.java
@@ -3,7 +3,6 @@ package app.freerouting.management;
 import app.freerouting.board.ItemIdentificationNumberGenerator;
 import app.freerouting.core.RoutingJob;
 import app.freerouting.io.FileFormat;
-import app.freerouting.interactive.HeadlessBoardManager;
 import app.freerouting.logger.FRLogger;
 
 /**
