+++ src/main/java/app/freerouting/management/RoutingJobScheduler.java
@@ -8,7 +8,6 @@ import app.freerouting.core.RoutingJobState;
 import app.freerouting.core.Session;
 import app.freerouting.core.StoppableThread;
 import app.freerouting.io.FileFormat;
-import app.freerouting.interactive.HeadlessBoardManager;
 import app.freerouting.logger.FRLogger;
 import app.freerouting.util.TextManager;
 import app.freerouting.util.gson.GsonProvider;
