+++ src/test/java/app/freerouting/drc/DesignRulesCheckerTest.java
@@ -12,7 +12,7 @@ import app.freerouting.management.BoardLoader;
 import app.freerouting.core.RoutingJob;
 import app.freerouting.core.Session;
 import app.freerouting.io.FileFormat;
-import app.freerouting.interactive.HeadlessBoardManager;
+import app.freerouting.management.HeadlessBoardManager;
 import app.freerouting.management.SessionManager;
 import app.freerouting.settings.GlobalSettings;
 import app.freerouting.fixtures.RoutingFixtureTest;
