+++ src/test/java/app/freerouting/interactive/InteractiveSettingsPropertyChangeTest.java
@@ -3,6 +3,7 @@ package app.freerouting.interactive;
 import app.freerouting.board.BoardObserverAdaptor;
 import app.freerouting.board.ItemIdentificationNumberGenerator;
 import app.freerouting.core.RoutingJob;
+import app.freerouting.management.HeadlessBoardManager;
 import app.freerouting.settings.RouterSettings;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
