+++ src/test/resources/archunit_store/9fcd8605-e576-4a10-9b13-9782f6985bca
@@ -1,11 +1,3 @@
-Constructor <app.freerouting.management.RoutingJobScheduler.<init>()> calls constructor <app.freerouting.interactive.HeadlessBoardManager.<init>(app.freerouting.core.RoutingJob)> in (RoutingJobScheduler.java:85)
-Constructor <app.freerouting.management.RoutingJobScheduler.<init>()> calls method <app.freerouting.interactive.HeadlessBoardManager.get_routing_board()> in (RoutingJobScheduler.java:93)
-Constructor <app.freerouting.management.RoutingJobScheduler.<init>()> calls method <app.freerouting.interactive.HeadlessBoardManager.loadFromKiCadJson(java.io.InputStream, app.freerouting.board.BoardObservers, app.freerouting.datastructures.IdentificationNumberGenerator)> in (RoutingJobScheduler.java:90)
-Constructor <app.freerouting.management.RoutingJobScheduler.<init>()> calls method <app.freerouting.interactive.HeadlessBoardManager.loadFromSpecctraDsn(java.io.InputStream, app.freerouting.board.BoardObservers, app.freerouting.datastructures.IdentificationNumberGenerator)> in (RoutingJobScheduler.java:87)
-Method <app.freerouting.api.v1.JobControllerV1.getDrcReport(java.lang.String)> calls constructor <app.freerouting.interactive.HeadlessBoardManager.<init>(app.freerouting.core.RoutingJob)> in (JobControllerV1.java:1195)
-Method <app.freerouting.api.v1.JobControllerV1.getDrcReport(java.lang.String)> calls method <app.freerouting.interactive.HeadlessBoardManager.get_routing_board()> in (JobControllerV1.java:1201)
-Method <app.freerouting.api.v1.JobControllerV1.getDrcReport(java.lang.String)> calls method <app.freerouting.interactive.HeadlessBoardManager.loadFromKiCadJson(java.io.InputStream, app.freerouting.board.BoardObservers, app.freerouting.datastructures.IdentificationNumberGenerator)> in (JobControllerV1.java:1197)
-Method <app.freerouting.api.v1.JobControllerV1.getDrcReport(java.lang.String)> calls method <app.freerouting.interactive.HeadlessBoardManager.loadFromSpecctraDsn(java.io.InputStream, app.freerouting.board.BoardObservers, app.freerouting.datastructures.IdentificationNumberGenerator)> in (JobControllerV1.java:1199)
 Method <app.freerouting.autoroute.BatchAutorouter.autoroute_pass(int)> calls constructor <app.freerouting.interactive.RatsNest.<init>(app.freerouting.board.BasicBoard)> in (BatchAutorouter.java:450)
 Method <app.freerouting.autoroute.BatchAutorouter.autoroute_pass(int)> calls constructor <app.freerouting.interactive.RatsNest.<init>(app.freerouting.board.BasicBoard)> in (BatchAutorouter.java:539)
 Method <app.freerouting.autoroute.BatchAutorouter.autoroute_pass(int)> calls constructor <app.freerouting.interactive.RatsNest.<init>(app.freerouting.board.BasicBoard)> in (BatchAutorouter.java:636)
@@ -36,12 +28,3 @@ Method <app.freerouting.autoroute.BatchOptimizer.opt_route_item(app.freerouting.
 Method <app.freerouting.core.RoutingJob.read_rules_file(java.lang.String, java.lang.String, java.lang.String, app.freerouting.interactive.GuiBoardManager, java.lang.String)> calls method <app.freerouting.interactive.GuiBoardManager.get_routing_board()> in (RoutingJob.java:161)
 Method <app.freerouting.core.RoutingJob.read_rules_file(java.lang.String, java.lang.String, java.lang.String, app.freerouting.interactive.GuiBoardManager, java.lang.String)> calls method <app.freerouting.interactive.GuiBoardManager.get_routing_board()> in (RoutingJob.java:168)
 Method <app.freerouting.core.RoutingJob.read_rules_file(java.lang.String, java.lang.String, java.lang.String, app.freerouting.interactive.GuiBoardManager, java.lang.String)> has parameter of type <app.freerouting.interactive.GuiBoardManager> in (RoutingJob.java:0)
-Method <app.freerouting.management.BoardLoader.loadBoardIfNeeded(app.freerouting.core.RoutingJob)> calls constructor <app.freerouting.interactive.HeadlessBoardManager.<init>(app.freerouting.core.RoutingJob)> in (BoardLoader.java:44)
-Method <app.freerouting.management.BoardLoader.loadBoardIfNeeded(app.freerouting.core.RoutingJob)> calls constructor <app.freerouting.interactive.HeadlessBoardManager.<init>(app.freerouting.core.RoutingJob)> in (BoardLoader.java:52)
-Method <app.freerouting.management.BoardLoader.loadBoardIfNeeded(app.freerouting.core.RoutingJob)> calls method <app.freerouting.interactive.HeadlessBoardManager.get_routing_board()> in (BoardLoader.java:49)
-Method <app.freerouting.management.BoardLoader.loadBoardIfNeeded(app.freerouting.core.RoutingJob)> calls method <app.freerouting.interactive.HeadlessBoardManager.get_routing_board()> in (BoardLoader.java:57)
-Method <app.freerouting.management.BoardLoader.loadBoardIfNeeded(app.freerouting.core.RoutingJob)> calls method <app.freerouting.interactive.HeadlessBoardManager.loadFromKiCadJson(java.io.InputStream, app.freerouting.board.BoardObservers, app.freerouting.datastructures.IdentificationNumberGenerator)> in (BoardLoader.java:45)
-Method <app.freerouting.management.BoardLoader.loadBoardIfNeeded(app.freerouting.core.RoutingJob)> calls method <app.freerouting.interactive.HeadlessBoardManager.loadFromSpecctraDsn(java.io.InputStream, app.freerouting.board.BoardObservers, app.freerouting.datastructures.IdentificationNumberGenerator)> in (BoardLoader.java:53)
-Method <app.freerouting.management.RoutingJobSchedulerActionThread.setJobOutput(app.freerouting.core.RoutingJob)> calls constructor <app.freerouting.interactive.HeadlessBoardManager.<init>(app.freerouting.core.RoutingJob)> in (RoutingJobSchedulerActionThread.java:281)
-Method <app.freerouting.management.RoutingJobSchedulerActionThread.setJobOutput(app.freerouting.core.RoutingJob)> calls method <app.freerouting.interactive.HeadlessBoardManager.replaceRoutingBoard(app.freerouting.board.RoutingBoard)> in (RoutingJobSchedulerActionThread.java:282)
-Method <app.freerouting.management.RoutingJobSchedulerActionThread.setJobOutput(app.freerouting.core.RoutingJob)> calls method <app.freerouting.interactive.HeadlessBoardManager.saveAsSpecctraSessionSes(java.io.OutputStream, java.lang.String)> in (RoutingJobSchedulerActionThread.java:286)
